% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elmr-data.R
\docType{data}
\name{dense_core_neurons}
\alias{dense_core_neurons}
\title{Sample EM tracings (homologous pair of neurons wrapping around lateral horn)}
\description{
Sample EM tracings (homologous pair of neurons wrapping around lateral horn)
}
\details{
These neurons were traced by Greg Jefferis because they appeared to be a
homologous pair from the left and right side of the brain. They contain dense
core vesicles (which made it easier to find matching cell bodies) and skirt
around the outside of the lateral horn.
}
\examples{
plot(dense_core_neurons)
\dontrun{
dense_core_neurons <- read.neurons.catmaid(c(713968, 822008))
}
}

