% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngraph.R
\name{distal_to}
\alias{distal_to}
\title{Return indices of points in a neuron distal to a given node}
\usage{
distal_to(x, node.idx = NULL, node.pointno = NULL, root = rootpoints(x),
  ...)
}
\arguments{
\item{x}{A neuron}

\item{node.idx, node.pointno}{The id of node beyond which distal points will
be selected. \code{node.idx} defines the integer index (counting from 1)
into the neuron's point array.}

\item{root}{The root node of the neuron for the purpose of selection.}

\item{...}{Additional}
}
\description{
Return indices of points in a neuron distal to a given node
}
