% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elmr-data.R
\docType{data}
\name{elm.landmarks}
\alias{elm.landmarks}
\alias{elm.landmarks.12}
\alias{elm.landmarks.13}
\title{ELM Landmarks created by Bock & Saalfeld groups at Janelia}
\description{
These are used to map locations in EM (FAFB) space to JFRC2013 (aka DPX)
space, in conjunction with a thin plate spline library.

\code{elm.landmarks.12} contains the landmark positions used for
  the FAFB12 assembly as a historical record.

\code{elm.landmarks.13} contains the landmark positions used for
  the FAFB13 assembly (from March to Sep 2017) as a historical record.
}
\details{
X,Y,Z correspond to locations in JFRC2013 template brain. However,
  they are in raw voxel coordinates. In contrast, X1,Y1,Z1 (which contain
  locations in FAFB space) are in nm.

  The current iteration of these landmarks are in FAFB13 space
}
\examples{
head(elm.landmarks)

\dontrun{
# Show landmarks in the context of the JFRC2013 template brain
library(nat.flybrains)
plot3d(JFRC2013)
# note that the landmarks are in raw voxel coordinates and must be scaled
xyz=scale(elm.landmarks[,c("X","Y","Z")], scale = 1/voxdims(JFRC2013), center = FALSE)
spheres3d(xyz, col = ifelse(elm.landmarks$Use, "green", 'red'), radius = 4)
}
}
\references{
See \url{https://github.com/saalfeldlab/elm} for the original source of the
landmarks.
}
