% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{elmr-defunct}
\alias{elmr-defunct}
\title{Functions that have been removed or moved to other packages}
\description{
Functions that have been removed or moved to other packages
}
\section{moved}{

\itemize{

\item \code{\link[nat]{distal_to}} has been moved to the \code{\link{nat}} package.

}
}

