% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elmr-package.R
\docType{package}
\name{elmr-package}
\alias{elmr}
\alias{elmr-package}
\title{Bridge light and EM flybrain datasets}
\description{
\bold{elmr} provides builds on the \bold{nat} and \bold{catmaid} packages to
provide tools to read, analyse, plot, transform and convert neuroanatomical
data, especially as it relates to whole brain EM voumes generated at the HHMI
Janelia Research Campus (groups of Davi Bock, Stephan Saalfeld and many
collaborators).
}
\keyword{package}

