% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nblast_fafb.R
\name{fetchn_fafb}
\alias{fetchn_fafb}
\alias{fetchdp_fafb}
\title{Fetch neurons from FAFB CATMAID server, transforming them as appropriate}
\usage{
fetchn_fafb(skids, mirror = TRUE, conn = NULL,
  reference = nat.flybrains::FCWB)

fetchdp_fafb(skids, mirror = TRUE, conn = NULL, ...)
}
\arguments{
\item{skids}{catmaid skeleton ids (see \code{\link[catmaid]{catmaid_skids}})
or a neuronlist}

\item{mirror}{whether to mirror the neuron (default \code{TRUE} since
FlyCircuit neurons are on fly's left and most FAFB tracings are on fly's
right.)}

\item{conn}{a \code{catmaid} connection object (see
\code{\link[catmaid]{catmaid_connection}})}

\item{reference}{The reference brain to which the neurons will be
transformed, either a \code{\link[nat.templatebrains]{templatebrain}}
object such as \code{\link[nat.flybrains]{FCWB}} or a character vector
naming such an object.}

\item{...}{Additional parameters passed to \code{\link[nat.nblast]{nblast}}}
}
\description{
\code{fetchn_fafb} is a thin wrapper around the
  \code{catmaid::\link[catmaid]{read.neurons.catmaid}} function with the
  additional convenience of automating conversion to a reference brain.
  Input can also be a pre-downloaded neuronlist.

\code{fetchdp_fafb} extends \code{fetchn_fafb} by additionally converting
  neurons to the \code{\link[nat]{dotprops}} representation suitable for
  \code{\link[nat.nblast]{nblast}}.
}
\examples{
\dontrun{
# fetch all neurons with an annotation exactly matching "KC" and convert
# to JFRC2013 space
kcs=fetchn_fafb("annotation:^KC$", reference=JFRC2013)

# supplying a neuronlist, it will mirror and convert the neurons
# to a reference brain
dc_nl=fetchn_fafb(dense_core_neurons, reference=JFRC2013)
}
}
