% Generated by roxygen2 (4.0.1): do not edit by hand
\name{inv_logit}
\alias{inv_logit}
\title{Inverse Logit Transformation}
\usage{
inv_logit(alpha)
}
\arguments{
\item{#alpha}{is a vector of real number}
}
\value{
a vector of inverse logit transformed values
   if(alpha>705) return(1) #otherwise too big to exp()
}
\description{
Perform the inverse logit transformation for data with zeros and ones included #' (Phil Dixon's suggestion)
}
\examples{
pc=seq(0,1,by=0.01)
cbind(pc, inv_logit(logit(pc)))
plot(pc, logit(pc), pch=19, col="gray", cex=1.2, ylab="transformed value")
points(pc, log(pc/(1-pc)), pch=20)
plot(pc, inv_logit(logit(pc)))
}
\seealso{
logit
}
\keyword{inverse}
\keyword{logit}
\keyword{tranform}

