% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logit}
\alias{logit}
\title{Logit Transformation}
\usage{
logit(x)
}
\arguments{
\item{x}{is a vector of values between 0 and 1 inclusive}
}
\value{
a vector of logit transformed values
}
\description{
Perform a logit transformation for data with zeros and ones included
(Phil Dixon's suggestion)
}
\examples{
pc=seq(0,1,by=0.01)
plot(pc, logit(pc), pch=19, col="gray", cex=1.2, ylab="transformed value")
points(pc, log(pc/(1-pc)), pch=20)
logit()
}
\seealso{
inv_logit
}
\keyword{logit}
\keyword{tranform}

