% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logit.transform}
\alias{logit.transform}
\title{Logit Transformation}
\usage{
logit.transform(x)
}
\arguments{
\item{x}{is a vector of values between 0 and 1 inclusive}
}
\description{
Perform a logit transformation for data with zeros and ones included
(Phil Dixon's suggestion)
}
\examples{
logit.transform(seq(0,1))
}
\keyword{logit}
\keyword{tranform}

