% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logw0}
\alias{logw0}
\title{Log transformation with zeros}
\usage{
logw0(x)
}
\arguments{
\item{x}{is a vector of values >=0}
}
\value{
preserves orders of magnitudes and results in 0 when original data is zero.
}
\description{
Perform a log transformation on a vector including zeros
}
\details{
posmin(x) is the smallest non-zero value in the data
   trunc(x) truncates x to an interger by dropping digits after decimal
   c=int(log(posmin(x)))
   d=exp(c)
   logw0(x)=log(x+d)-c
}
\examples{
tx=seq(0,5,by=0.001)
plot(tx, log(tx))
plot(tx, log(tx+1))
plot(tx, logw0(tx))
pairs(cbind(orig=tx, log=log(tx), logp1=log(tx+1), logw0=logw0(tx) ), #' lower.panel=NULL, row1attop = FALSE)
Note logp1 changes overall shape of transform- the log relationship is lost.
}
\references{
McCune, Bruce and James Grace. 2002. Analysis of Ecological Communities. MjM Software Design, Glenden Beach OR. pgs 68-69
}
\seealso{
logit
}
\keyword{log}
\keyword{transform}

