% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_fafb.R
\name{open_fafb}
\alias{open_fafb}
\title{Generate URLs to open FAFB CATMAID in browser at a given XYZ location}
\usage{
open_fafb(x, s = rgl::select3d(), mirror = FALSE,
  sample = elmr::FAFB, rowwise = NA, open = interactive() & !rowwise,
  zoom = 1, active_skeleton_id = NULL, active_node_id = NULL,
  conn = NULL, ...)
}
\arguments{
\item{x}{A numeric vector or any object compatible with
\code{\link[nat]{xyzmatrix}} (see details)}

\item{s}{Optional selection function of the type returned by
\code{\link[rgl]{select3d}}}

\item{mirror}{Whether to mirror the point to the opposite side of the brain}

\item{sample}{The template brain space associated with the coordinates in
\code{x}}

\item{rowwise}{When \code{rowwise=TRUE} each point in \code{x} is converted
to a separate URL.}

\item{open}{Whether to open the url in the browser or simply return it.
Defaults to \code{TRUE} when R is running in interactive mode.}

\item{zoom}{The CATMAID zoom factor (defaults to 1)}

\item{active_skeleton_id, active_node_id}{Set highlighted skeleton and node in
CATMAID.}

\item{conn}{An optional CATMAID connection that specifies the server URL
(otherwise a harcoded URL will be used).}

\item{...}{Additional arguments to be added to URL.}
}
\description{
Generate URLs to open FAFB CATMAID in browser at a given XYZ location
}
\details{
Note that the default behaviour is that if object \code{x} contains
  exactly one point then a single URL will be generated and by default
  CATMAID will be opened immediately at that location. If there is more than
  1 point, by default the function will stop and wait for the user to make an
  interactive selection in a \code{rgl} window.

  You can control this behaviour by using the \code{open} and \code{rowwise}
  arguments. When \code{open=FALSE}, then the URLs will be returned but not
  opened in a web browser. If \code{rowwise=TRUE} (i.e. multiple URLs are
  being generated) then \code{open} will be set to \code{FALSE}.

  When \code{rowwise=TRUE} each point in \code{x} is converted to a separate
  URL. As a special case, if \code{x} is a data.frame generated by
  \code{catmaid_get_treenodes_detail} then URLs will be generated for every
  row (i.e. each treenode). See examples for details.
}
\examples{
open_fafb(c(316, 143, 26), sample=JFRC2013, open=FALSE)
library(nat)
\dontrun{
open3d()
plot3d(kcs20)
# waits for user to draw a selection rectangle
open_fafb(kcs20, sample=FCWB)
# same but mirrors selected points to opposite hemisphere
open_fafb(kcs20, sample=FCWB, mirror=TRUE)

# Uses last CATMAID connection to specify URL to open
open_fafb(kcs20, sample=FCWB, conn=catmaid_login())
}

\donttest{
# make one URL for each row of the input
open_fafb(matrix(runif(n=6, max=1000), ncol = 3), rowwise=TRUE)
}

\donttest{
## Special case using catmaid_get_treenodes_detail() as input
# fetch a neuron and extract the node ids for a given tag
n=read.neuron.catmaid(16)
tagged = n$tags[['TODO']]
tagged_details <- catmaid_get_treenodes_detail(tagged)
# now find FAFB URLs
tagged_details$url <- open_fafb(tagged_details)
# take a look at results
head(tagged_details)
}

\dontrun{
# copy results to clipboard e.g. to paste into a spreadsheet
library(clipr)
write_clip(tagged_details)
}
}
\seealso{
\code{\link{xform_brain}},
  \code{\link{catmaid_get_treenodes_detail}}
}
