% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{OSelm_train.formula}
\alias{OSelm_train.formula}
\title{Trains an extreme learning machine with random weights}
\usage{
OSelm_train.formula(formula, data, Elm_type, nHiddenNeurons, ActivationFunction,
  N0, Block)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted.}

\item{data}{training data frame containing the variables specified in formula.}

\item{Elm_type}{select if the ELM must perform a "regression" or "classification"}

\item{nHiddenNeurons}{number of neurons in the hidden layer}

\item{ActivationFunction}{"rbf" for radial basis function with Gaussian kernels , "sig" for sigmoidal fucntion, "sin" for sine function, "hardlim" for hard limit function}

\item{N0}{size of the first block to be processed}

\item{Block}{size of each chunk to be processed at each step}
}
\value{
returns all the parameters used in the function, the weight matrix, the labels for the classification, the number of classes found, the bias, the beta activation function and the accuracy on the trainingset
}
\description{
Trains an extreme learning machine with random weights
}

