% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{preProcess}
\alias{preProcess}
\title{Pre processing function for the training and test data set. Each numeric variable is standardized between -1 and 1 and each categorical variable is coded with a dummy coding.}
\usage{
preProcess(data)
}
\arguments{
\item{data}{to be preprocesses}
}
\value{
return the pre processed dataset
}
\description{
Pre processing function for the training and test data set. Each numeric variable is standardized between -1 and 1 and each categorical variable is coded with a dummy coding.
}

