% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{read.neurons.fafb}
\alias{read.neurons.fafb}
\alias{fafb_get_meta}
\title{Read neurons from CATMAID with extra meta-data}
\usage{
read.neurons.fafb(
  skids,
  meta = c("neuron name", "whimsical name", "cell type", "ItoLee_Lineage",
    "ItoLee_Hemilineage", "Hartenstein_Lineage", "Hartenstein_Hemilineage",
    "Putative_Neurotransmitter", "tracing status", "hemibrain match", "paper"),
  sub = ".*:",
  OmitFailures = TRUE,
  ...
)

fafb_get_meta(
  skids,
  meta = c("neuron name", "whimsical name", "cell type", "ItoLee_Lineage",
    "ItoLee_Hemilineage", "Hartenstein_Lineage", "Hartenstein_Hemilineage",
    "Putative_Neurotransmitter", "tracing status", "hemibrain match", "paper"),
  sub = ".*:",
  OmitFailures = TRUE,
  ...
)
}
\arguments{
\item{skids}{One or more numeric skeleton ids or a character vector defining
a query (see \code{\link[catmaid]{catmaid_skids}} or examples for the syntax).}

\item{meta}{a vector of meta-annotations to query.
The annotations labelled by these meta-annotations will appear as
entries in the \code{data.frame} attached to the returned \code{neuronlist}.}

\item{sub}{what to remove from pulled annotations, to clean them up a little.
Can be set to "" if you do not wish to remove anything.}

\item{OmitFailures}{Whether to omit neurons for which \code{FUN} gives an
error. The default value (\code{NA}) will result in \code{\link{nlapply}}
stopping with an error message the moment there is an error. For other
values, see details.}

\item{...}{Additional arguments passed to the \code{\link[catmaid]{catmaid_fetch}}
function}
}
\description{
Use \code{read.neurons.catmaid} to acquire neurons from a CATMAID instance.
In addition, include as meta-data certain annotations that these neurons have. To select which
annotations appear as entries in the neurons' meta-data, select 'meta-annotations' to appear as the
data field. By default, meta-annotations for developmental lineage identities are used,
to get the lineage / hemilineage to which a FAFB Drosophila neuron belongs.
}
\examples{
\donttest{
\dontrun{
# Get some information on a published neuron
## 11519370 was pubished first in Dolan et al. 2019
meta = fafb_get_meta("11519370")

# We can get all of the lineage annotations made in the FAFB dataset
## These come from two differebnt naming schemes, one by Volker Hartenstein
## and another by the groups' of K. Ito and T. Lee.
fafb.neurons.lineage =
read.neurons.fafb("annotation:Lineage_annotated")
}}
}
