% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{xform_brain}
\alias{xform_brain}
\title{Transform data between template brains including FAFB}
\usage{
xform_brain(x, sample, reference, ...)
}
\arguments{
\item{x}{Neurons or other data}

\item{sample}{Source template brain (e.g. IS2) that data is currently in.}

\item{reference}{Target template brain (e.g. IS2) that data should be
transformed into.}

\item{...}{extra arguments to pass to \code{\link[nat]{xform}}.}
}
\description{
this function overrides the nat.templatebrains function of the
  same name. It will be removed when nat.templatebrains/nat provide support
  for additional types of registration data.

wraps \code{\link[nat.templatebrains]{xform_brain}}
}
\examples{
library(nat)
kcs13.fafb=xform_brain(kcs20[1:3], sample=FCWB, reference=FAFB11)
}

