% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngraph.R
\name{distal_to}
\alias{distal_to}
\title{Return indices of points in a neuron distal to a given node}
\usage{
distal_to(x, node.idx = NULL, node.pointno = NULL, root = rootpoints(x),
  ...)
}
\arguments{
\item{x}{A neuron}

\item{node.idx, node.pointno}{The id of node beyond which distal points will
be selected. \code{node.idx} defines the integer index (counting from 1)
into the neuron's point array.}

\item{root}{The root node of the neuron for the purpose of selection.}

\item{...}{Additional}
}
\description{
Return indices of points in a neuron distal to a given node
}
\examples{
\donttest{
## Fetch a finished DL1 projection neuron
finished_pns=catmaid_get_neuronnames('annotation:^LH_DONE')
# should only be one neuron but pick first just in case
dl1skid=names(grep('DL1', finished_pns, value = TRUE))[1]
dl1=read.neuron.catmaid(dl1skid)

## subset to part of neuron distal to a tag "SCHLEGEL_LH"
# nb distal_to can accept either the PointNo vertex id or raw index as a
# pivot point
dl1.lh=subset(dl1, distal_to(dl1,node.pointno = dl1$tags$SCHLEGEL_LH))
plot(dl1,col='blue', WithNodes = FALSE)
plot(dl1.lh, col='red', WithNodes = FALSE, add=TRUE)
}
}
\seealso{
\code{\link[nat]{subset.neuron}}, \code{\link[nat]{prune}}
}
