% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elmr-data.R
\docType{data}
\name{elm.landmarks}
\alias{elm.landmarks}
\title{ELM Landmarks created by Bock & Saalfeld groups at Janelia}
\description{
These are used to map locations in EM (FAFB) space to JFRC2013 (aka DPX)
space, in conjunction with a thin plate spline library.
}
\examples{
head(elm.landmarks)
}

