% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\docType{package}
\name{elmr-defunct}
\alias{elmr-defunct}
\alias{elmr-defunct-package}
\title{Defunct function(s) in the elmr package}
\usage{
"elmr-defunct"()
}
\description{
These functions are removed from the elmr package.
}
\section{Details}{

\tabular{rl}{
  \code{distal_to} \tab is moved to \code{\link[nat]{distal_to}}\cr
  \code{simplify_neuron} \tab is moved to \code{\link[nat]{simplify_neuron}}\cr
  \code{stitch_neuron} \tab is moved to \code{\link[nat]{stitch_neuron}}\cr
  \code{stitch_neurons} \tab is moved to \code{\link[nat]{stitch_neurons}}\cr
  \code{prune_twigs} \tab is moved to \code{\link[nat]{prune_twigs}}\cr
  \code{tpsreg} \tab is moved to \code{\link[nat]{tpsreg}}\cr
}
}

