% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{elmr-defunct}
\alias{elmr-defunct}
\title{Functions that have been removed or moved to other packages}
\description{
Functions that have been removed or moved to other packages
}
\section{moved}{
 \itemize{

  \item \code{\link[nat]{distal_to}} has been moved to the \code{\link{nat}}
  package.

  \item \code{\link[nat]{simplify_neuron}} has been moved to the
  \code{\link{nat}} package.

  \item \code{\link[nat]{stitch_neuron}} has been moved to the
  \code{\link{nat}} package.

  \item \code{\link[nat]{stitch_neurons}} has been moved to the
  \code{\link{nat}} package.

  \item \code{\link[nat]{prune_twigs}} has been moved to the
  \code{\link{nat}} package.

  \item \code{\link[nat]{tpsreg}} has been moved to the \code{\link{nat}}
  package.

  \item \code{\link[nat]{plot3d.ngraph}} has been moved to the
  \code{\link{nat}} package.

  }
}

