% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elmr-data.R
\docType{data}
\name{FAFB}
\alias{FAFB}
\alias{FAFB11}
\alias{FAFB12}
\alias{FAFB13}
\title{templatebrain for assembled FAFB full adult female brain EM volume}
\description{
\bold{FAFB} \code{\link[nat.templatebrains]{templatebrain}}
  objects encapsulate the physical dimensions of assembled FAFB serial
  section data. There is a different \code{templatebrain} for each assembly
  version.They are used to define the transformations between FAFB and other
  templates like \code{\link[nat.flybrains]{JFRC2013}} (see
  \code{\link{xform_brain}})

\code{FAFB13} Assembly v13 of FAFB full adult female brain EM
  volume (2016-10).

\code{FAFB12} Assembly v12 of FAFB full adult female brain EM
  volume (2016-04-01)

\code{FAFB11} Assembly v11 of FAFB full adult female brain EM
  volume (2016-02-09)
}
\details{
\code{FAFB12} is the default FAFB assembly since 2016-04-01.
  \code{FAFB13} is in preview as of 2016-10-20 but is not production ready.
  Furthermore the dimensions of the FAFB13 volume are not yet confirmed. It
  is hoped that the v13 assembly will correct some substantial distortions in
  anterior parts of the brain such as the antennal lobe and initial
  indications are promising.
}
\examples{
FAFB12
dim(FAFB12)
voxdims(FAFB12)
boundingbox(FAFB12)
}
\seealso{
\code{\link[nat.templatebrains]{templatebrain}},
  \code{\link{xform_brain}}
}

