% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elmr-data.R
\docType{data}
\name{FAFB13.surf}
\alias{FAFB13.surf}
\alias{FAFB13NP.surf}
\title{FAFB13 brain and neuropil surface objects}
\description{
\code{FAFB13.surf} describes the neuropil surface of the FAFB13
  brain.

\code{FAFB13NP.surf} is a surface object for the Ito et al
  (2014) neuropil regions in FAFB13 space
}
\details{
These neuropil regions are generated from the
  \code{\link[nat.flybrains]{JFRC2.surf}} and
  \code{\link[nat.flybrains]{JFRC2NP.surf}} objects, respectively. Because
  they are not defined from FAFB13 itself by transformed using bridging
  registrations, they are only as good as those bridging registrations. In
  particular although the \code{\link{elm.landmarks}} used to bridge
  FAFB13-JFRC2013 are well defined in the fly's right brain hemisphere, they
  are much less dense in other areas. Therefore these surfaces are clearly
  not less accurate in the left brain hemisphere and show significanct
  displacements in the optic lobes.

  The acronyms for the neuropil regions are those used in Ito et al. 2014 and
  can also be looked up at \url{http://virtualflybrain.org}.
}
\examples{
boundingbox(FAFB13.surf)
FAFB13NP.surf$RegionList
\dontrun{
plot3d(FAFB13.surf)
# a short cut which plots the surface in semi-transparent grey.
clear3d()
plot3d(FAFB13)

# plot right lateral horn
plot3d(FAFB13NP.surf, "LH_R", alpha=0.4)
# plot both antennal lobe surfaces (second arg is regex if no exact match)
# note use of initial anchor (^) to avoid matches to LAL
plot3d(FAFB13NP.surf, "^AL", alpha=0.4)
}
}
\seealso{
\code{\link{FAFB13}}, \code{\link{FAFB13}}
}

