% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{fafb_hemilineage_contents}
\alias{fafb_hemilineage_contents}
\title{Create list of FAFB neurons within a hemilineage}
\usage{
fafb_hemilineage_contents(
  hemilineage,
  side = c("right", "left"),
  nomenclature = c("ItoLee", "Hartenstein"),
  path = getwd(),
  return = c("data.frame", "csv"),
  ...
)
}
\arguments{
\item{hemilineage}{a valid hemilineage name. In CATMAID FAFB v14, these appear as annotation that start either in the form \code{ItoLee_Hemilineage}
or \code{hartenstein_Hemilineage}.}

\item{side}{the side of the brain for which you want to get neurons. Depends on the annotation \code{"side: left"}.}

\item{nomenclature}{whether you are looking for hemilineages in the form \code{ItoLee_Hemilineage}
or \code{Hartenstein_Hemilineage}.}

\item{path}{where to save the outputted \code{.csv} file.}

\item{return}{whether to return a \code{data.frame} or write to a \code{.csv} at location \code{path}.}

\item{...}{methods passed to \code{catmaid} functions.}
}
\description{
This functions creates a \code{.csv} file listing every neuron annotated as belonging to a certain hemilineage
within the FAFB volume. A secondary hemilineage is a set of neurons born together in the late larval stage.
They form ~90% of the neurons in FAFB.

To see the available hemilineages you can load the package \code{hemibrainr} and see \code{hemibrain_hemilineages}.
}
\examples{
\donttest{
\dontrun{
## A hemilineage of mostly laterla horn neurons
df = fafb_hemilineage_contents("DL2_dorsal", side = "right")

# from a list of found hemilineages
gs = googlesheets4::read_sheet("1HI8RZJhV8aWC6hw5T0qh2
__9D8V0zKdtxx2kkDsuI8Y")
for(i in 1:nrow(gs)){
  hl = gs[i,"ItoLee_Hemilineage"]
  message("Processing ", hl)
  fafb_hemilineage_contents(hl, side = "right", return = "csv")
  fafb_hemilineage_contents(hl, side = "left", return = "csv")
}
}}
}
