% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nblast_fafb.R
\name{nblast_fafb}
\alias{nblast_fafb}
\title{NBLAST EM tracing against flycircuit (or other databases of) neurons}
\usage{
nblast_fafb(skids, db = NULL, conn = NULL, mirror = TRUE,
  normalised = TRUE, reverse = FALSE, reference = nat.flybrains::FCWB,
  .parallel = TRUE, ...)
}
\arguments{
\item{skids}{catmaid skeleton ids (see \code{\link[catmaid]{catmaid_skids}})
or a neuronlist}

\item{db}{A neuronlist object containing neurons to search. Defaults to the
value of \code{options('nat.default.neuronlist')}.}

\item{conn}{a \code{catmaid} connection object (see
\code{\link[catmaid]{catmaid_connection}})}

\item{mirror}{whether to mirror the neuron (default \code{TRUE} since
flycircuit neurons are on fly's left and most FAFB tracings are on fly's
right.)}

\item{normalised}{Whether to return normalised NBLAST scores}

\item{reverse}{Treat the FAFB skeleton as NBLAST target rather than query
(sensible if \code{db} contains partial skeletons/tracts; default
\code{FALSE}).}

\item{reference}{The reference brain to which the neurons will be
transformed, either a \code{\link[nat.templatebrains]{templatebrain}}
object such as \code{\link[nat.flybrains]{FCWB}} or a character vector
naming such an object.}

\item{.parallel}{Whether to parallelise the nblast search (see details and
also \code{\link[nat.nblast]{nblast}} for how to use the parallel interface
provided by the \code{doMC} package.)}

\item{...}{Additional parameters passed to \code{\link[nat.nblast]{nblast}}}
}
\value{
an object of class nblastfafb for which \code{plot3d}, \code{summary}
  and \code{hist} methods exist.
}
\description{
NBLAST EM tracing against flycircuit (or other databases of) neurons
}
\details{
Still depends on having a \code{\link[nat]{neuronlist}} containing
  registered neurons (usually from flycircuit.tw). The example code downloads
  a set of projection neurons. The full data must be requested from Greg
  Jefferis.

  When \code{.parallel=TRUE}, nblast_fafb will parallelise the search across
  multiple cores if possible. You can set an option specifying a default
  number of cores \code{elmr.nblast.cores}. See \code{\link{elmr-package}}.
  Otherwise the default will either be roughly half the number of cores as
  determined by \code{\link[doParallel]{registerDoParallel}}. If your machine
  does not have multiple cores (or you do not want to use them), the search
  will be a bit more efficient if you set the option
  \code{elmr.nblast.cores=1} or the argument \code{.parallel=TRUE}.
}
\examples{
\dontrun{
# first load neuronlist object containing registered neurons (see details)
allpndps=flycircuit::load_si_data('allpndps.rds')
# ... and set that as the default for queries and plotting
options(nat.default.neuronlist='allpndps')

# then make sure you are loged in to catmaid server
# catmaid::catmaid_login(<your connection args>)

# nblast neuron 27884
PN27884f=nblast_fafb(27884, mirror = FALSE)
# summary table of results
summary(PN27884f)
# plot results, just top hit
plot3d(PN27884f, hits=1)

# alternatively, you can supply an existing neuronlist
dc_nl = nblast_fafb(dense_core_neurons, mirror=TRUE)
summary(dc_nl)
}
}
\seealso{
\code{\link[nat.nblast]{nblast}},
  \code{\link[doParallel]{registerDoParallel}} for setting spreading load
  across cores.
}
