% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nblast_fafb.R
\name{nblast_fafb}
\alias{nblast_fafb}
\title{NBLAST EM tracing against flycircuit neurons}
\usage{
nblast_fafb(skids, db = NULL, conn = NULL, mirror = TRUE,
  normalised = TRUE, .parallel = T, ...)
}
\arguments{
\item{skids}{catmaid skeleton ids (see \code{\link[catmaid]{catmaid_skids}})}

\item{db}{A neuronlist object containing neurons to search}

\item{conn}{a \code{catmaid} connection object (see
\code{\link[catmaid]{catmaid_connection}})}

\item{mirror}{whether to mirror the neuron (default \code{TRUE} since
flycircuit neurons are on fly's left and most FAFB tracings are on fly's
right.)}

\item{normalised}{Whether to return normalised NBLAST scores}

\item{.parallel}{Whether to parallelise the nblast search (see details of
\code{\link[nat.nblast]{nblast}} for how to use the parallel interface
provided by the \code{doMC} package.)}

\item{...}{Additional parameters passed to \code{\link[nat.nblast]{nblast}}}
}
\value{
an object of class nblastfafb for which \code{plot3d}, \code{summary}
  and \code{hist} methods exist.
}
\description{
NBLAST EM tracing against flycircuit neurons
}
\details{
Still depends on an object called \code{dps} containing registered
  flycircuit.tw neurons that must be obtained by contacting Greg Jefferis.
}
\examples{
\dontrun{
# first load dps object (see details)

# nblast neuron 27884
PN27884f=nblast_fafb(27884, mirror = FALSE)
# summary table of results
summary(PN27884f)
# plot results, just top hit
plot3d(PN27884f, hits=1)
}
}
\seealso{
\code{\link[nat.nblast]{nblast}}
}

