% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_fafb.R
\name{open_fafb}
\alias{open_fafb}
\title{Open FAFB CATMAID in browser at a given XYZ location}
\usage{
open_fafb(x, s = rgl::select3d(), mirror = FALSE, sample = elmr::FAFB13,
  zoom = 1, open = interactive(), active_skeleton_id = NULL,
  active_node_id = NULL, ...)
}
\arguments{
\item{x}{A numeric vector or any object compatible with
\code{\link[nat]{xyzmatrix}} (see details)}

\item{s}{A selection function of the type returned by
\code{\link[rgl]{select3d}}}

\item{mirror}{Whether to mirror the point to the opposite side of the brain}

\item{sample}{The template brain space associated with the coordinates in
\code{x}}

\item{zoom}{The CATMAID zoom factor (defaults to 1)}

\item{open}{Whether to open the url in the browser or simply return it.
Defaults to \code{TRUE} when R is running in interactive mode.}

\item{active_skeleton_id, active_node_id}{Set highlighted skeleton and node in
CATMAID.}

\item{...}{Additional arguments to be added to URL.}
}
\description{
Open FAFB CATMAID in browser at a given XYZ location
}
\details{
Note that if object \code{x} contains exactly one point then CATMAID
  will be opened immediately at that location, whereas if there is more than
  1 point, the function will stop and wait for the user to make an
  interactive selection in a \code{rgl} window.
}
\examples{
open_fafb(c(316, 143, 26), sample=JFRC2013, open=FALSE)
library(nat)
\dontrun{
open3d()
plot3d(kcs20)
# waits for used to draw a selection rectangle
open_fafb(kcs20, sample=FCWB)
# same but mirrors selected points to opposite hemisphere
open_fafb(kcs20, sample=FCWB, mirror=TRUE)
}
}
\seealso{
xform_brain
}

