% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_neuron.R
\name{simplify_neuron}
\alias{simplify_neuron}
\title{Simplify a neuron to include its root and two longest branches}
\usage{
simplify_neuron(x, ...)
}
\arguments{
\item{x}{A \code{\link[nat]{neuron}} to simplify}

\item{...}{Additional arguments (currently ignored)}
}
\value{
The simplified \code{neuron} or the untouched original neuron for
  neurons that have <=1 branch point.
}
\description{
Simplify a neuron to include its root and two longest branches
}
\details{
This is still a somewhat experimental function (hence its location
  in elmr rather than nat). It would be nice to add the option to retain the
  most important N branches, rather than the first 1 as here.
}
\examples{
\donttest{
dl1=read.neuron.catmaid(catmaid_skids('name:DL1')[1])
dl1.simp=simplify_neuron(dl1)
plot(dl1, col='green', WithNodes = F)
plot(dl1.simp, col='red', add = T)
}
}
\seealso{
\code{\link[nat]{spine}}
}
