% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.R
\name{summarise_contribution}
\alias{summarise_contribution}
\alias{suggest_authorship}
\alias{write_authors}
\alias{write_ack}
\title{Summarise tracer contributions / suggest authorship for set of neurons}
\usage{
summarise_contribution(skids, type = c("nodes", "synapses", "pre",
  "post"), ...)

suggest_authorship(x, auth = 5, ackn = c(3000, 100))

write_authors(x, ...)

write_ack(x, ...)
}
\arguments{
\item{skids}{catmaid skeleton ids (see \code{\link[catmaid]{catmaid_skids}})
or a neuronlist}

\item{type}{Whether to summarise contributions towards arbour (nodes) or
synapses}

\item{...}{Additional arguments passed by \code{summarise_contribution} to
\code{\link[catmaid]{catmaid_get_contributor_stats}} or by
\code{write_authors/write_ack} to \code{suggest_authorship}}

\item{x}{A data frame containing the output of summarise_contribution}

\item{auth}{The minimum percentage of the total effort for authorship}

\item{ackn}{The minimum number of nodes/synapses to be acknowledged. When
length of 2, the first number is assumed to be the criterion for arbour
nodes, the second for synapses.}
}
\value{
A data.frame describing contributions or for
  \code{write_authors/write_ack} a string describing in descending order of
  contribution written to the console by \code{\link{cat}}.
}
\description{
\code{summarise_contribution} produces a data.frame of
  contributions by user. \code{suggest_authorship} can use this to suggest
  which contributors meet defined authorship/acknowledgement criteria.

\code{suggest_authorship} suggests how to acknowledge
  individuals

\code{write_authors} produces a list of authors

\code{write_ack} produces an acknowledgements wording.
}
\examples{
\donttest{
pnsc=summarise_contribution("annotation:^PN$")
suggest_authorship(pnsc)
write_authors(pnsc, auth=3.0)
write_ack(pnsc)

pnsc=summarise_contribution("annotation:^PN$", type='synapses')
}
}
