% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xform_tps.R
\name{tpsreg}
\alias{tpsreg}
\alias{xformpoints.tpsreg}
\title{Thin plate spline registrations via nat::xform and friends}
\usage{
tpsreg(sample, reference, ...)

\method{xformpoints}{tpsreg}(reg, points, swap = NULL, ...)
}
\arguments{
\item{sample, reference}{Matrices defining the sample (or floating) and
reference (desired target after transformation) spaces. See details.}

\item{...}{additional arguments passed to xformpoints.tpsreg}

\item{reg}{The \code{tpsreg} registration object}

\item{points}{The 3D points to transform}

\item{swap}{Whether to change the direction of registration (default of
\code{NULL} checks if reg has a \code{attr('swap'=TRUE)}) otherwise}
}
\description{
\code{tpsreg} creates an object encapsulating a thin plate spine
  transform mapping a paired landmark set.

\code{xformpoints.tpsreg} enables \code{\link[nat]{xform}} and
  friends to transform 3d vertices (or more complex objects containing 3d
  vertices) using a thin plate spline mapping stored in a \code{tpsreg}
  object.
}
\details{
Note that we use the \bold{nat} convention for naming the
  sample/reference space arguments but these actually clash with the
  nomenclature in the underlying \code{Morpho::\link[Morpho]{tps3d}}
  function. \itemize{

  \item refmat (Morpho3d) == sample (nat)

  \item tarmat (Morpho3d) == reference (nat)

  }
}
\seealso{
\code{\link[nat]{reglist}}, \code{\link[nat]{read.landmarks}}
}

