% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xform_tps.R
\name{tpsreg}
\alias{tpsreg}
\alias{xformpoints.tpsreg}
\title{Thin plate spline registrations via nat::xform and friends}
\usage{
tpsreg(refmat, tarmat, ...)

\method{xformpoints}{tpsreg}(reg, points, swap = FALSE, ...)
}
\arguments{
\item{refmat, tarmat}{The reference and target matrices}

\item{...}{additional arguments passed to xformpoints.tpsreg}

\item{reg}{The \code{tpsreg} registration object}

\item{points}{The 3D points to transform}

\item{swap}{Whether to change the direction of registration}
}
\description{
\code{tpsreg} creates an object encapsulating a thin plate spine
  transform mapping a paired landmark set.

\code{xformpoints.tpsreg} provides
}
\seealso{
\code{\link{reglist}}, \code{\link[nat]{read.landmarks}}
}

