% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{xform_brain}
\alias{xform_brain}
\title{Transform data between template brains including FAFB}
\usage{
xform_brain(x, sample, reference, ...)
}
\arguments{
\item{x}{Neurons or other data}

\item{sample}{Source template brain (e.g. IS2) that data is currently in.}

\item{reference}{Target template brain (e.g. IS2) that data should be
transformed into.}

\item{...}{extra arguments to pass to \code{\link[nat]{xform}}.}
}
\description{
this function overrides the nat.templatebrains function of the
  same name. It will be removed when nat.templatebrains/nat provide support
  for additional types of registration data.

wraps \code{\link[nat.templatebrains]{xform_brain}}
}
\examples{
# position of antennal lobe glomeruli "V" in JFRC2013 template brain
vgloms.jfrc2013=data.frame(X=c(316,229),
  Y=c(143, 139),
  Z=c(26,22),
  row.names=c("V_L", "V_R"))
# Convert to FAFB12 coordinates
xform_brain(vgloms.jfrc2013, sample = JFRC2013, reference = FAFB12)

\dontrun{
# Conversion of neurons from FlyCircuit template
# NB this conversion depends on a full install of nat.flybrains and CMTK
library(nat)
kcs13.fafb=xform_brain(kcs20[1:3], sample=FCWB, reference=FAFB12)
}
}

